LayoutTests/http/wpt/cross-origin-options/allow-postmessage.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Tests that postMessage() works when 'Cross-Origin-Options: allow-postmessage' HTTP header is served</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/utils.js"></script>
</head>
<body>
<script>

promise_test(function(test) {
    return withIframe("cross-origin-options-allow-postmessage-pong.html", true /* isCrossOrigin */).then((f) => {
        return new Promise((resolve) => {
            window.onmessage = (msg) => {
                window.onmessage = null;
                assert_equals(msg.data, "PONG");
                assert_equals(msg.source, f.contentWindow);
                resolve();
            };
            assert_throws("SecurityError", function() { f.contentWindow.length }, "length property access");
            f.contentWindow.postMessage("PING", "*");
        });
    });
}, "postMessage() on Cross-origin iframe with 'Cross-Origin-Options: allow-postmessage' HTTP header");

promise_test(function(test) {
    return withPopup("cross-origin-options-allow-postmessage-pong.html", true /* isCrossOrigin */).then((result) => {
        return new Promise((resolve) => {
            window.onmessage = (msg) => {
                window.onmessage = null;
                assert_equals(msg.data, "PONG");
                assert_equals(msg.source, result.window);
                resolve();
            };
            assert_throws("SecurityError", function() { result.window.length }, "length property access");
            result.window.postMessage("PING", "*");
        });
    });
}, "postMessage() on Cross-origin popup with 'Cross-Origin-Options: allow-postmessage' HTTP header");

</script>
</body>
</html>
